Moved loading to loadingicon in parts
